Extracted ENS nat-lab tests to separate file
